Added a 32-bit bootstrapping Assembly file
Going to use the boot sector to enter protected mode (32-bit), then load
and jump into the second sector, which will setup the GDT for and then
switch the chip into 64-bit long mode. After this, the following sectors
will contain the kernel so we will load the kernel and call `kmain`.