Switch back from CMake to GNU Make
CMake was causing lots of problems with linker path resolution which
could be fixed by using a more simple build system. The project is not
very large at the moment and I have a better understanding of what GNU
Make is doing so I can debug it easier. Also, my understanding of
standard makefiles was lacking, so this is good practice.

This could become an issue further down the line, but I will cross that
bridge if and when I get to it.